StudentLiveActivity: Don't keep screen enabled
While it makes sense for the lecturer activity because the lecturer uses
the phone only in a passive way, the student can turn on and off the
screen as he directly interacts with the device.

This helps to save power in case the student forgot to turn off the
screen.